Fixed number formatting

+++ src/main/java/io/github/cshadd/fetch_bot/Core.java
@@ -104,7 +104,7 @@ implements FetchBot {
                 arduinoComm.pullRobot();
 
                 // Sensors
-                final int ultrasonicSensor = Integer.parseInt(arduinoComm.getRobotValue("s"));
+                final int ultrasonicSensor = (int)Float.parseFloat(arduinoComm.getRobotValue("s"));
                 if (ultrasonicSensor != -1) {
                     if (ultrasonicSensor != currentUltrasonicSensor) {
                         currentUltrasonicSensor = ultrasonicSensor;
